\name{perms}
\alias{perms}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Permutations}
\description{
Returns all possible permutations given the number of elements.
}
\usage{
perms(n)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{n}{ Number of elements. }
}

\value{
  {Returns a matrix containing all possible permutations.}
}

\author{ Volodymyr Melnykov, Wei-Chen Chen, and Ranjan Maitra. }

\seealso{ \code{ClassProp}. }
\examples{
%\dontrun{
perms(3)
%}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ cluster }

